a stoplight draws the tier its reading is in
A falling band is closed at its better end, so a reading on the rung comes out
with a whole band travelled, and the arc drawn for that covered the ring and
repainted it in the tier ahead: the Tasks tile read blue at a reading of 1 while
the status line read yellow off the same payload. An arc filling the ring is no
longer drawn, so the track's own tier shows.

The upkeep hours are written as whole numbers. The figure text is one size for
every stoplight of a scale, so a figure that will not fit the circle has to be
shorter rather than smaller, and `-3.26` did not fit.

+++ readout-system/readouts/pages/upkeep-capacity/upkeep-capacity.readout.ts
@@ -10,7 +10,7 @@ export const upkeepCapacity = {
   label: "Capacity",
   unit: "hours",
   place: 3,
-  figureFormat: "decimal",
+  figureFormat: "integer",
   scaleSlug: "capacity-hours",
   groupSlugs: ["upkeep"],
   wireKey: "capacity",
