+++ readout-system/readouts/pages/upkeep-sleep/upkeep-sleep.readout.ts
@@ -10,7 +10,7 @@ export const upkeepSleep = {
   label: "Sleep",
   unit: "hours",
   place: 6,
-  figureFormat: "decimal",
+  figureFormat: "integer",
   scaleSlug: "sleep-hours",
   groupSlugs: ["upkeep"],
   wireKey: "sleep",
