+++ readout-system/readouts/pages/upkeep-surplus/upkeep-surplus.readout.ts
@@ -10,7 +10,7 @@ export const upkeepSurplus = {
   label: "Surplus",
   unit: "hours",
   place: 2,
-  figureFormat: "decimal",
+  figureFormat: "integer",
   scaleSlug: "surplus-hours",
   groupSlugs: ["upkeep", "surplus"],
   wireKey: "surplus",
